onyx/ai/pyproject.toml
[tool.poetry]
authors = ["Hai Tran"]
name = "onyx-ai"
version = "0.0.0"
description = ""
readme = "README.md"
packages = [{ include = "onyx", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
pydantic = { extras = ["email"], version = "^2.6.4" }
langchain = "^0.2.0"
langchain-core = "^0.2.0"
langchain-community = "^0.2.0"
langchain-openai = "^0.1.8"
openai = "^1.33.0"
clickhouse-connect = "0.7.8"
tiktoken = "^0.7.0"
langfuse = "^2.39.1"
pyvespa = "^0.44.0"
lark = "^1.1.9"
dspy-ai = "^2.4.12"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
